Fixes to compile, strip and pass tests on Android (CppTests that is, C4Tests are still a mess)
Note: /data/local/tmp is not a valid long term solution (but should be good enough for testing) for a temporary directory.  Will need to create an APK for running unit tests probably...

Needed to change selectFrom to take a string stream by reference instead of returning one because otherwise it can't compile with GCC 4.x stdc++ (Can't use libc++ on Android because it crashes every time an exception is thrown)

+++ LiteCore/Support/Logging.cc
@@ -93,7 +93,7 @@ namespace litecore {
                 strcat(source, " ");
                 strcat(source, name);
             }
-            __android_log_write(kLevels[level], source, message);
+            __android_log_write(kLevels[(int)level], source, message);
         #else
             #if __APPLE__
                 struct timeval tv;
